refactoring
shortened some methods, removed commented out code, hard coded component ID for testing SLADIM-264
