+++ src/issue-creator/issue-creator.ts
@@ -3,6 +3,7 @@ import { HttpService } from '@nestjs/common';
 import { LogMessageFormat } from 'logging-format';
 import { IssueFormat } from '../IssueFormat';
 import { ConfigService } from '@nestjs/config';
+import { Logs } from 'src/schema/logs.schema';
 /**
  * generic IssueCreator class. It extends IssueReporter to report issues that were created.
  * It implements the IssueCreatorComponent to handle the incoming logs.
@@ -21,19 +22,32 @@ export abstract class IssueCreator extends IssueReporter {
    */
   async createIssueFromLog(log: LogMessageFormat) {
     const issue: IssueFormat = {
-      title: `${log.type}`,
+      title: `${log.type} + Error`,
       body: `${log.data}`,
       category: 'BUG',
-      componentIDs: [`${log.detectorUrl}`, `${log.sourceUrl}`],
-      labels: [`${log.detectorUrl}`],
-      assignees: [`${log.detectorUrl}`],
-      locations: [`${log.sourceUrl}`],
+      componentIDs: [`5d31793fbdabf003`],
       startDate: log.time,
-      clientMutationID: 'id1234',
+      clientMutationID: 'Error-Monitoring',
     };
     return this.reportIssue(issue);
   }
 
+  async checkIssueID(query: Logs[], log: LogMessageFormat){
+    const relatedLog = query.find(log => log.issueID);
+    if (relatedLog) {
+      console.log('FOUND', relatedLog);
+      if (!relatedLog.issueID) {
+        // Issue already exists but latest log doesn't have a IssueId, this should not happen but if it does we create a new issue anyways
+        console.log('WARNING: Log does not have a IssueId');
+        return await this.createIssueFromLog(log);
+      }
+      console.log('Updating Issue with Id');
+      return  await this.updateLastOccurrence(relatedLog.issueID, log.time); 
+    } else {
+      console.log('Issue does not exist yet');
+      return await this.createIssueFromLog(log);
+    }
+  }
   /**
    * basic functionality to handle logs that every IssueCreator should have
    *
